New textures for Skin settings and radio buttons.
Also cleaned up the scripts and changed how the synopsis script works, I now pass the video to the skin so synopsis doesn't need the video preview settings.
(ie, selecting a extension, I now look for one and if found send it to the skin)

Also updated the Skin settings & added synopsis mode to the Preview Video Settings screen.

Added a Yes/No dialog to the Clean Thumbnails script & an OK dialog so the user can pick what they want to do.
(has a description on what its going to do)

Cleaned up the Includes_Video_Preview.xml as some buttons aren't needed anymore.

Also the default video preview window only works with Games\Game name\preview.xmv, wmv, mp4 or mpg so simplify stuff.

+++ Mod Files/scripts/XBMC4Kids/Utilities/CleanThumbs.py
@@ -2,6 +2,9 @@
 '''
 	Script by Rocky5
 	Removed all the .tbn/dds file inside the "thumbnail\programs\" directory for said profile.
+	
+	Update: 28 July 2016
+	-- Added a Yes No dialog and a OK dialog at the end, also cleaned it up a bit.
 '''
 ########################################################################################################################################
 
@@ -15,6 +18,9 @@ import time
 
 ThumbPath			= xbmc.translatePath( "special://profile/thumbnails/programs" )
 Sub_Directories		= [ "0", "1", "2", "3", "4", "5", "6", "7", "8", "9", "0", "a", "b", "c", "d", "e", "f", "fanart" ]
+pDialog = xbmcgui.DialogProgress()
+dialog = xbmcgui.Dialog()
+
 
 ########################################################################################################################################
 # Start markings for the log file.
@@ -23,33 +29,34 @@ print "=========================================================================
 print "| Scripts\XBMC4Kids\Utilities\CleanThumbs.py loaded."
 print "| ------------------------------------------------------------------------------"
 
+if dialog.yesno('Thumbnail Cleaner','','Remove all the [B]Game[/B] Thumbsnails?.','They will be reloaded when you enter the games section.') == 1:
+	pDialog.create('Cleaning Thumbnails')
+	pDialog.update(0, 'Please wait')
+	time.sleep(0.1)
+	pDialog.update(20, 'Please wait')
+	time.sleep(0.1)
+	pDialog.update(50, 'Please wait')
 
-pDialog = xbmcgui.DialogProgress()
-pDialog.create('Cleaning Thumbnails')
-pDialog.update(0, 'Please wait')
-time.sleep(0.2)
-pDialog.update(20, 'Please wait')
-time.sleep(0.2)
-pDialog.update(50, 'Please wait')
-time.sleep(0.2)
-pDialog.update(55, 'Please wait')
-time.sleep(0.2)
-pDialog.update(80, 'Please wait')
+	for Sub_Directories in Sub_Directories :
+		Thumbs_Directory = os.path.join( ThumbPath, Sub_Directories )
+		if os.path.isdir( Thumbs_Directory ) :
+			TBNFiles = glob.glob( os.path.join( Thumbs_Directory, "*.tbn" ) )
+			for tbn in TBNFiles:
+				os.remove(tbn)
+				DDSFiles = glob.glob( os.path.join( Thumbs_Directory, "*.dds" ) )
+				for dds in DDSFiles:
+					os.remove(dds)
 
-for Sub_Directories in Sub_Directories :
-	Thumbs_Directory = os.path.join( ThumbPath, Sub_Directories )
-	if os.path.isdir( Thumbs_Directory ) :
-		TBNFiles = glob.glob( os.path.join( Thumbs_Directory, "*.tbn" ) )
-		DDSFiles = glob.glob( os.path.join( Thumbs_Directory, "*.dds" ) )
-		for tbn in TBNFiles:
-			#print tbn
-			os.remove(tbn)
-			for dds in DDSFiles:
-				#print tbn
-				os.remove(dds)
+	pDialog.update(55, 'Please wait')
+	time.sleep(0.1)
+	pDialog.update(80, 'Please wait')
 
-print '| Removed all .tbn/dds files from "' + ThumbPath + '"'
-print "================================================================================"
+	print '| Removed all .tbn/dds files from "' + ThumbPath + '"'
+	print "================================================================================"
 
-pDialog.update(100, 'Program Thumbnails cleaned.')
-time.sleep(1.0)	
+	pDialog.update(100, 'Please wait')
+	pDialog.close()
+	dialog.ok('Thumbnail Cleaner','','All Thumbnails Removed.')
+else:
+	print '| You picked no.'
+	print "================================================================================"
